Separated playermovementbehaviiour
