+++ Assets/_BForBoss/_Core/Scripts/Character/PlayerMovementBehaviour.IGetPlayerTransform.cs
@@ -1,0 +1,10 @@
+using Perigon.Weapons;
+using UnityEngine;
+
+namespace BForBoss
+{
+    public partial class PlayerMovementBehaviour : IGetPlayerTransform
+    {
+        Transform IGetPlayerTransform.Value => rootPivot;
+    }
+}
