+++ Assets/_BForBoss/_Core/Scripts/Character/PlayerMovementBehaviour.IGetPlayerTransform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e841a60f3c7f09c498659557b28c3afa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
