+++ Assets/_BForBoss/_Core/Scripts/Character/PlayerMovementBehaviour.IPlayerMovementStates.cs
@@ -1,6 +1,3 @@
-using Perigon.Weapons;
-using UnityEngine;
-
 namespace BForBoss
 {
     public interface IPlayerMovementStates
@@ -27,25 +24,5 @@ namespace BForBoss
 
         public bool IsWallRunning => _wallRunBehaviour.IsWallRunning;
         public bool IsRunning => IsWalking();
-    }
-    
-    public partial class PlayerMovementBehaviour : IGetPlayerTransform
-    {
-        Transform IGetPlayerTransform.Value => rootPivot;
-    }
-
-    public partial class PlayerMovementBehaviour : IWeaponBobIntensity
-    {
-        float IWeaponBobIntensity.Value
-        {
-            get
-            {
-                var canBob = (_wallRunBehaviour.IsWallRunning || IsOnGround()) && 
-                             !_dashBehaviour.IsDashing && 
-                             !_slideBehaviour.IsSliding && 
-                             GetVelocity().magnitude > 0;
-                return GetVelocity().magnitude * (canBob ? 1 : 0);
-            }
-        }
     }
 }
