+++ Assets/_BForBoss/_Core/Scripts/Character/PlayerMovementBehaviour.IPlayerMovementStates.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2136a34da59a02748b735981f58ea97a
+guid: d28be641b90933a4595b13047ae8fbd5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
