+++ Assets/_BForBoss/_Core/Scripts/Character/PlayerMovementBehaviour.IWeaponBobIntensity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8b9e6f60f6f0d747bdc2657bbc42650
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
